feat: simplify backgrounds and unify status messages
